combine lc_names and lc_subjects functionality into shared library_of_congress module
