+++ inference/app/src/aggregate.py
@@ -2,7 +2,7 @@ import logging
 from copy import deepcopy
 
 from .exact.lc_names import get_lc_names_data
-from .exact.lc_subjects import get_lc_subjects_data
+from .exact.library_of_congress import get_lc_subjects_data, get_lc_names_data
 from .exact.mesh import get_mesh_data
 from .exact.utils import (loc_id_to_wikidata_id, mesh_id_to_wikidata_id,
                           wikidata_id_to_alt_source_ids)
